+++ teamcity/resource_agent_pool_project_attachment.go
@@ -1,0 +1,84 @@
+package teamcity
+
+import (
+	"github.com/Cardfree/teamcity-sdk-go/teamcity"
+	"github.com/Cardfree/teamcity-sdk-go/types"
+	"github.com/hashicorp/terraform/helper/schema"
+
+	"fmt"
+	"log"
+)
+
+func resourceAgentPoolProjectAttachment() *schema.Resource {
+	return &schema.Resource{
+		Create: resourceAgentPoolProjectAttachementCreate,
+		Read:   resourceAgentPoolProjectAttachementRead,
+		Delete: resourceAgentPoolProjectAttachementDelete,
+
+		Schema: map[string]*schema.Schema{
+			"pool": &schema.Schema{
+				Type:     schema.TypeString,
+				Required: true,
+				ForceNew: true,
+			},
+			"project": &schema.Schema{
+				Type:     schema.TypeString,
+				Required: true,
+				ForceNew: true,
+			},
+		},
+	}
+}
+
+func resourceAgentPoolProjectAttachementCreate(d *schema.ResourceData, meta interface{}) error {
+	client := meta.(*teamcity.Client)
+
+	attachmentPool := d.Get("pool").(string)
+	attachmentProject := d.Get("project").(string)
+
+	attachment := types.AgentPoolAttachment{
+		ProjectID: attachmentProject,
+	}
+
+	pool, pool_err := client.GetAgentPool(d.Get("pool").(string))
+	if pool_err != nil {
+		return pool_err
+	}
+
+	create_err := client.CreateAgentPoolProjectAttachment(attachmentPool, &attachment)
+	if create_err != nil {
+		return create_err
+	}
+
+	id := fmt.Sprintf("%d_%s", pool.ID, attachment.ProjectID)
+	d.SetId(id)
+	return nil
+}
+
+func resourceAgentPoolProjectAttachementRead(d *schema.ResourceData, meta interface{}) error {
+	log.Printf("Reading agent_pool_project_attachment resource %q", d.Id())
+	client := meta.(*teamcity.Client)
+	pool, err := client.GetAgentPool(d.Get("pool").(string))
+	if err != nil {
+		return err
+	}
+
+	log.Printf("[DEBUG] Read Agent Pools: %q", pool)
+	var project types.Project
+	project = pool.Projects[d.Get("project").(string)]
+
+	log.Printf("[DEBUG] Read Agent Pool: %q", project)
+
+	if project.ID == "" {
+		d.SetId("")
+		return nil
+	}
+
+	return nil
+}
+
+func resourceAgentPoolProjectAttachementDelete(d *schema.ResourceData, meta interface{}) error {
+	client := meta.(*teamcity.Client)
+	err := client.DeleteAgentPoolProjectAttachement(d.Get("pool").(string), d.Get("project").(string))
+	return err
+}